added categorybitmask, collisionbitmask and contactbitmask to the nodes

+++ THUtil.h
@@ -10,11 +10,11 @@
 
 static const int THProjectileSpeed = 400;
 
-typedef NS_OPTIONS(NSUInteger, THCollisionCategory) {
-    THCollisionCategoryEnemy      = 1 << 0,
-    THCollisionCategoryProjectile = 1 << 1,
-    THCollisionCategoryDebris     = 1 << 2,
-    THCollisionCategoryGround     = 1 << 3,
+typedef NS_OPTIONS(uint32_t, THCollisionCategory) {
+    THCollisionCategoryEnemy      = 1 << 0,       //  0000
+    THCollisionCategoryProjectile = 1 << 1,       //  0010
+    THCollisionCategoryDebris     = 1 << 2,       //  0100
+    THCollisionCategoryGround     = 1 << 3        //  1000
 };
 
 @interface THUtil : NSObject
